chore: Renamed marketing-page folder to home
